codes/Publisher/requirements.txt
charm-crypto
gpiozero
psutil
pycryptodome
requests